backend/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
bcrypt
defusedxml
python-multipart
pydantic
pytest
pytest-asyncio
httpx

